% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Datasets.R
\docType{data}
\name{Birdstack}
\alias{Birdstack}
\title{Raster Stack with five variables measured in Pohnpei Island in Micronesia}
\format{A raster stack with 5 layers:}
\usage{
Birdstack
}
\description{
A Raster Stack containing five rasters from in the  Pohnpei Island in Micronesia
where the detection histories of the IslandBirds where taken.
\describe{
  \item{Elev}{Metres above sea level of the sampled point}
  \item{AgroFo}{Proportion of a station (from 0 to 1) corresponding to
  agricultural forest incudes areas withsubsistence or commercial staple
  crops mixed with undisturbed forest}
  \item{SecVec}{Proportion of a station (from 0 to 1) corresponding to
  secondary vegetation which represents anthropogenic habitats (e.g.,
  savannah) and early colonizers and invasive species (e.g. \emph{Hibiscus spp.}
   and \emph{Merremia peltata)}}
  \item{Wetland}{Proportion of a station (from 0 to 1) corresponding to areas
  classified as mangroves, marshes, swamp forest and fresh water bodies}
  \item{Upland}{Proportion of a station (from 0 to 1) corresponding to upland
  and undisturbed forest habitats}
}
}
\seealso{
\code{\link[DiversityOccupancy]{IslandBirds}}

\code{\link[DiversityOccupancy]{Daily_Cov}}

\code{\link[DiversityOccupancy]{siteCov}}
}
\keyword{datasets}
