% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Datasets.R
\docType{data}
\name{Daily_Cov}
\alias{Daily_Cov}
\title{Seven covariables measured at 120 survey stations in an island in micronesia}
\format{A list with 6 elements, each of them with a data frame}
\usage{
Daily_Cov
}
\description{
A list containing the measurements of seven variables correspondant to 120
different survey stations (Rows) in Pohnpei Island where the detection
histories of the IslandBirds where taken.
}
\details{
\describe{
  \item{Day}{Ordinal day}
  \item{Wind}{Wind intensity using a modified Beafourt scale}
  \item{Obs}{First initial of the observer}
  \item{Time}{Survey time (minutes after sunrise)}
  \item{Rain}{Categorical variable where 0 = no rain, 1 = light rain and
  2 = heavy rain}
  \item{Noise}{Ambient noise (1-10 being 10 the loudest)}
  \item{Clouds}{Cloud cover in percentage}
}
}
\references{
Oleiro P. 2014. Avian population responses to anthropogenic
landscape changes in Pohnpei, Federated Stats of Micronesia. Masters Thesis
University of Missouri.
}
\seealso{
\code{\link[DiversityOccupancy]{IslandBirds}}

\code{\link[DiversityOccupancy]{siteCov}}
}
\keyword{datasets}
