% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Datasets.R
\docType{data}
\name{IslandBirds}
\alias{IslandBirds}
\title{Occupancy data of 5 bird species in Micronesia.}
\format{A data frame with 120 rows and 20 variables}
\usage{
IslandBirds
}
\description{
A dataset containing the detections history of 5 species in the
Pohnpei Island for 4 consecutive days (Columns) in 120 different sites (Rows).
\describe{
  \item{CICA.1}{Detection history of \emph{Coracina tenurostris}: on day 1}
  \item{CICA.2}{Detection history of \emph{Coracina tenurostris}: on day 2}
  \item{CICA.3}{Detection history of \emph{Coracina tenurostris}: on day 3}
  \item{CICA.4}{Detection history of \emph{Coracina tenurostris}: on day 4}
  \item{CIRW.1}{Detection history of \emph{Acrocephalus syrinx}: on day 1}
  \item{CIRW.2}{Detection history of \emph{Acrocephalus syrinx}: on day 2}
  \item{CIRW.3}{Detection history of \emph{Acrocephalus syrinx}: on day 3}
  \item{CIRW.4}{Detection history of \emph{Acrocephalus syrinx}: on day 4}
  \item{CIWE.1}{Detection history of \emph{Zosterops semperi}: on day 1}
  \item{CIWE.2}{Detection history of \emph{Zosterops semperi}: on day 2}
  \item{CIWE.3}{Detection history of \emph{Zosterops semperi}: on day 3}
  \item{CIWE.4}{Detection history of \emph{Zosterops semperi}: on day 4}
  \item{LBWE.1}{Detection history of \emph{Rukia longirostra}: on day 1}
  \item{LBWE.2}{Detection history of \emph{Rukia longirostra}: on day 2}
  \item{LBWE.3}{Detection history of \emph{Rukia longirostra}: on day 3}
  \item{LBWE.4}{Detection history of \emph{Rukia longirostra}: on day 4}
  \item{MIPI.1}{Detection history of \emph{Ducula oceanica}: on day 1}
  \item{MIPI.2}{Detection history of \emph{Ducula oceanica}: on day 2}
  \item{MIPI.3}{Detection history of \emph{Ducula oceanica}: on day 3}
  \item{MIPI.4}{Detection history of \emph{Ducula oceanica}: on day 4}
}
}
\references{
Oleiro P. 2014. Avian population responses to anthropogenic
landscape changes in Pohnpei, Federated Stats of Micronesia. Masters Thesis
University of Missouri.
}
\seealso{
\code{\link[DiversityOccupancy]{Daily_Cov}}

\code{\link[DiversityOccupancy]{siteCov}}
}
\keyword{datasets}
