% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Datasets.R
\docType{data}
\name{siteCov}
\alias{siteCov}
\title{Nine covariables measured at 120 survey stations in an island in micronesia}
\format{A data frame with 120 rows and 52 variables:}
\usage{
siteCov
}
\description{
A dataframe containing the measurements of 5 variables (columns) correspondant
to 120 different survey stations (Rows) in Pohnpei Island where the detection
histories of the IslandBirds where taken.
\describe{
  \item{Elev}{Metres above sea level of the sampled point}
  \item{AgroFo}{Proportion of a station (from 0 to 1) corresponding to
  agricultural forest incudes areas withsubsistence or commercial staple
  crops mixed with undisturbed forest}
  \item{SecVec}{Proportion of a station (from 0 to 1) corresponding to
  secondary vegetation which represents anthropogenic habitats (e.g.,
  savannah) and early colonizers and invasive species (e.g. \emph{Hibiscus spp.}
   and \emph{Merremia peltata)}}
  \item{Wetland}{Proportion of a station (from 0 to 1) corresponding to areas
  classified as mangroves, marshes, swamp forest and fresh water bodies}
  \item{Upland}{Proportion of a station (from 0 to 1) corresponding to upland
  and undisturbed forest habitats}
}
}
\references{
Oleiro P. 2014. Avian population responses to anthropogenic
landscape changes in Pohnpei, Federated Stats of Micronesia. Masters Thesis
University of Missouri.
}
\seealso{
\code{\link[DiversityOccupancy]{Daily_Cov}}

\code{\link[DiversityOccupancy]{IslandBirds}}
}
\keyword{datasets}
